test: update tests per openAuthoring changes, split tests into multiple files
