+++ cypress/integration/editorial_workflow_spec_github_backend_graphql.js
@@ -1,0 +1,5 @@
+import fixture from './github/editorial_workflow';
+
+describe('Github Backend Editorial Workflow - GraphQL API', () => {
+  fixture({ use_graphql: true });
+});
