+++ cypress/integration/editorial_workflow_spec_github_backend_rest.js
@@ -1,0 +1,5 @@
+import fixture from './github/editorial_workflow';
+
+describe('Github Backend Editorial Workflow - REST API', () => {
+  fixture({ use_graphql: false });
+});
