+++ cypress/integration/editorial_workflow_spec_github_backend_rest_open_authoring.js
@@ -1,0 +1,5 @@
+import fixture from './github/open_authoring';
+
+describe('Github Backend Editorial Workflow - REST API - Open Authoring', () => {
+  fixture({ use_graphql: false });
+});
